refact: project folder structure
