+++ .env
@@ -1,0 +1,2 @@
+REACT_APP_API_KEY = 8cac6dec66e09ab439c081b251304443
+REACT_APP_ENDPOINT = https://api.themoviedb.org/3
