+++ src/components/TrailerDialog/trailer-dialog.scss
@@ -1,0 +1,4 @@
+.no-trailer-msg {
+  padding: "30px";
+  color: "black";
+}
