+++ src/constants.js
@@ -1,7 +1,2 @@
-export const API_KEY = "8cac6dec66e09ab439c081b251304443";
-export const ENDPOINT = "https://api.themoviedb.org/3";
-export const ENDPOINT_DISCOVER =
-  ENDPOINT + "/discover/movie?api_key=" + API_KEY + "&sort_by=vote_count.desc";
-export const ENDPOINT_SEARCH = ENDPOINT + "/search/movie?api_key=" + API_KEY;
-export const ENDPOINT_MOVIE =
-  ENDPOINT + "/movie/507086?api_key=" + API_KEY + "&append_to_response=videos";
+export const ENDPOINT_DISCOVER = `${process.env.REACT_APP_ENDPOINT}/discover/movie?api_key=${process.env.REACT_APP_API_KEY}&sort_by=vote_count.desc`;
+export const ENDPOINT_SEARCH = `${process.env.REACT_APP_ENDPOINT}/search/movie?api_key=${process.env.REACT_APP_API_KEY}`;
